Refactor: remove _span_utils.py shim, inline helpers in _tracer.py
Move _apply_error_attributes, _apply_llm_finish_attributes, and
_maybe_emit_llm_event from the _span_utils.py compatibility shim
directly into _tracer.py. Rename _apply_error_attributes to _set_error
for clarity. Delete the now-unused _span_utils.py module.

No behavioral changes — the replacement functions produce identical span
attributes and events.
